+++ src/lib/connectors/index.ts
@@ -2,7 +2,7 @@ import { Connection, Engine, ExecutionResult } from "@/types";
 import mysql from "./mysql";
 import postgres from "./postgres";
 import mssql from "./mssql";
-import { Schema } from "@/types/schema";
+import { Schema } from "@/types";
 
 export interface Connector {
   testConnection: () => Promise<boolean>;
