+++ src/lib/connectors/mssql/index.ts
@@ -1,7 +1,7 @@
 import { ConnectionPool } from "mssql";
 import { Connection, ExecutionResult } from "@/types";
 import { Connector } from "..";
-import { Schema } from "@/types/schema";
+import { Schema } from "@/types";
 
 const systemDatabases = ["master", "tempdb", "model", "msdb"];
