+++ src/types/database.ts
@@ -1,5 +1,4 @@
 import { Id } from ".";
-import { Schema } from "./schema";
 export interface Database {
   connectionId: Id;
   name: string;
@@ -11,4 +10,8 @@ export interface Table {
   // structure is a string of the table structure.
   // It's mainly used for providing a chat context for the assistant.
   structure: string;
+}
+export interface Schema {
+  name: string;
+  tables: Table[];
 }
