--- src/types/schema.ts
@@ -1,6 +1,0 @@
-import { Table } from ".";
-
-export interface Schema {
-  name: string;
-  tables: Table[];
-}
